fix: update return data structure
