+++ solagent-adapters/rig/dexscreener/Cargo.toml
@@ -1,16 +1,17 @@
 [package]
 name = "solagent-rig-dexscreener"
-version = "0.1.1"
+version = "0.1.2"
 edition = "2021"
 authors = ["zTgx <beautifularea@gmail.com>"]
 repository = "https://github.com/zTgx/solagent.rs"
 keywords = ["solagent", "rig", "dexscreener"]
 license = "Apache-2.0"
-description = "solagent.rs rig dexscreener"
+description = "solagent rig dexscreener"
 
 [dependencies]
-solagent-core = "0.1.3"
+solagent-core = "0.1.6"
+solagent-plugin-dexscreener = "0.1.2"
 solagent-parameters = "0.1.0"
 serde = { version = "1.0", features = ["derive"] }
-solagent-plugin-dexscreener = "0.1.1"
 thiserror = "2.0.9"
+serde_json = "1.0"
